server/requirements.txt
cherrypi
RPi.GPIO
